Converted editableElement to base class with managed deadkey info.
